minor: merge directory and entry repositories
