+++ src/StockportContentApi/Repositories/DirectoryRepository.cs
@@ -2,22 +2,26 @@
 
 namespace StockportContentApi.Repositories;
 
-public interface IDirectorypRepository
+public interface IDirectoryRepository
 {
     Task<HttpResponse> Get(string slug);
+    Task<HttpResponse> GetEntry(string slug);
 }
 
-public class DirectoryRepository : BaseRepository, IDirectorypRepository
+public class DirectoryRepository : BaseRepository, IDirectoryRepository
 {
     private readonly Contentful.Core.IContentfulClient _client;
     private readonly IContentfulFactory<ContentfulDirectory, Directory> _directoryFactory;
+    private readonly IContentfulFactory<ContentfulDirectoryEntry, DirectoryEntry> _directoryEntryFactory;
 
     public DirectoryRepository(ContentfulConfig config, 
                                 IContentfulClientManager clientManager,
-                                IContentfulFactory<ContentfulDirectory, Directory> directoryFactory)
+                                IContentfulFactory<ContentfulDirectory, Directory> directoryFactory,
+                                IContentfulFactory<ContentfulDirectoryEntry, DirectoryEntry> directoryEntryFactory)
     {
         _client = clientManager.GetClient(config);
         _directoryFactory = directoryFactory;
+        _directoryEntryFactory = directoryEntryFactory;
     }
 
     public async Task<HttpResponse> Get(string slug)
@@ -26,10 +30,34 @@ public class DirectoryRepository : BaseRepository, IDirectorypRepository
         var entries = await GetAllEntriesAsync(_client, builder);
         var contentfulDirectories = entries as IEnumerable<ContentfulDirectory> ?? entries.ToList();
 
-        var directories = contentfulDirectories.Select(g => _directoryFactory.ToModel(g));
+        var directory = contentfulDirectories.Select(g => _directoryFactory.ToModel(g)).SingleOrDefault();
+        directory.Entries = await GetEntriesInDirectory(directory.ContentfulId);
 
-        return entries == null || !directories.Any()
+        return entries is null || directory is null
             ? HttpResponse.Failure(HttpStatusCode.NotFound, $"Directory not found {slug}")
-            : HttpResponse.Successful(directories.ToList());
+            : HttpResponse.Successful(directory);
+    }
+    public async Task<HttpResponse> GetEntry(string slug)
+    {
+        var builder = new QueryBuilder<ContentfulDirectoryEntry>().ContentTypeIs("group").FieldEquals("fields.slug", slug).Include(2);
+        var entries = await GetAllEntriesAsync(_client, builder);
+        var contentfulDirectoryEntries = entries as IEnumerable<ContentfulDirectoryEntry> ?? entries.ToList();
+
+        var directoryEntry = contentfulDirectoryEntries.Select(g => _directoryEntryFactory.ToModel(g));
+
+        return entries == null || !directoryEntry.Any()
+            ? HttpResponse.Failure(HttpStatusCode.NotFound, $"Directory entry not found {slug}")
+            : HttpResponse.Successful(directoryEntry.ToList());
+    }
+
+    private async Task<IEnumerable<Model.DirectoryEntry>> GetEntriesInDirectory(string id)
+    {
+        var builder = new QueryBuilder<ContentfulDirectoryEntry>().ContentTypeIs("group").LinksToEntry(id).Include(2);
+        var entries = await GetAllEntriesAsync(_client, builder);
+        var contentfulDirectoryEntries = entries as IEnumerable<ContentfulDirectoryEntry> ?? entries.ToList();
+
+        var directoryEntries = contentfulDirectoryEntries.Select(g => _directoryEntryFactory.ToModel(g));
+
+        return directoryEntries;
     }
 }
